move the moc index tests to a separate file
